requirements.txt
# Django
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0

# Database
psycopg-binary==3.2.13
psycopg2==2.9.11
sqlalchemy==2.0.36

# Celery & Task Queue
celery[sqs]==5.4.0
kombu[sqs]==5.4.2
boto3==1.35.93
botocore==1.35.93
pycurl==7.45.3

# Validation & Data
pydantic==2.12.5
pydantic-core==2.41.5
email-validator==2.3.0

# Authentication & Security
PyJWT==2.10.1
cryptography==46.0.3

# Utilities
python-dotenv==1.2.1
requests==2.32.5
dnspython==2.8.0
beautifulsoup4==4.14.3
soupsieve==2.8.1

# JavaScript Engine (for EJS rendering)
py-mini-racer==0.6.0

# Type Checking
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0

# Other
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0

# Google
google-auth==2.38.1
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.2.0
google-api-python-client==2.169.0
pycparser==2.23

